bluetooth: move the declarations out of driver/
